# public/api-docs/swaggerDoc.json
{
    "swagger": "2.0",
    "info": {
        "description": "#### WeTravel\n #### WeTravel is an application that will enable its Company Staff book their international travel and accommodation globally; easily and conveniently across all the locations/centers where the Company has its operations. \n\n #### While early travel tended to be slower, more dangerous, and more dominated by trade and migration, WeTravel uses cultural and technological advances to make travel become easier and more accessible.",
        "version": "1",
        "title": "WeTravel",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "vokeolomu@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Admin create a new user",
                "description": "Allows an admin to create a new user",
                "operationId": "CreateUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@company.com"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
...
        "/api/v1/auth/remember-info": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "user remember info",
                "description": "Allows a user to remember his info for future requests",
                "operationId": "rememberInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "rememberInfo": {
                                    "type": "string",
                                    "example": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "token": {
                    "type": "integer",
                    "format": "int64"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@company.com"
                },
                "password": {
                    "type": "string"
                },
                "address": {
                    "type": "string",
                    "description": "Address of WeTravel User"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
        "Success": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "description": "success"
                }
            },
            "xml": {
                "name": "Success"
            }
        },
...
    }
}